requirements.txt
flask
requests
confluent_kafka==2.0.2
configparser
python-dotenv